Move PresetEditor to backend module
